Modif Racine Proj

+++ js/cart.js
@@ -5,19 +5,19 @@ const emptyCartAlert = `<div class="alert alert-warning alert-dismissible fade s
 </button>
 </div>`
 
-function NewCartRow(product) {
+function NewCartRow(product, i) {
     let totalPricePerProduct = product.quantity * product.price
     const cartRow = `<tr>
                         <th scope="row">${product.name}</th>
-                        <td><input class="quantityInput selectNumber" type="number" name="numberOfAddedProducts"
-                        value="${product.quantity}" min="1"></td>
+                        <td><input class="quantityInput" type="number" name="numberOfAddedProducts 
+                        onchange="updatequantity(this, ${i})" value="${product.quantity}" min="1"></td>
                         <td class="totalPricePerProduct">${totalPricePerProduct} €</td>
                         <td><button type="button" class="remove_product_btn btn btn-danger btn-sm"><svg width="15" aria-hidden="true" focusable="false" data-prefix="far" data-icon="trash-alt" class="svg-inline--fa fa-trash-alt fa-w-14" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M268 416h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12zM432 80h-82.41l-34-56.7A48 48 0 0 0 274.41 0H173.59a48 48 0 0 0-41.16 23.3L98.41 80H16A16 16 0 0 0 0 96v16a16 16 0 0 0 16 16h16v336a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V128h16a16 16 0 0 0 16-16V96a16 16 0 0 0-16-16zM171.84 50.91A6 6 0 0 1 177 48h94a6 6 0 0 1 5.15 2.91L293.61 80H154.39zM368 464H80V128h288zm-212-48h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12z"></path></svg></button></td>
                     <tr>`
     document.getElementById(`cartRowsContainer`).innerHTML += cartRow
 }
 
-//console.log(localStorage.getItem(`items`))
+//console.log(localStorage.getItem(`cart-key`))
 
 function GetItemsFromLocalStorage() {
     if (localStorage.getItem(`cart-key`) == null) {
@@ -26,12 +26,22 @@ function GetItemsFromLocalStorage() {
         cartContent = JSON.parse(localStorage.getItem('cart-key'))
         //console.log(cartContent)
         for (let i = 0; i < cartContent.length; i++) {
-            NewCartRow(cartContent[i])
+            NewCartRow(cartContent[i], i)
         }
     }
 }
 
 GetItemsFromLocalStorage()
+
+// changement de la quantité dans le panier
+
+function updatequantity(element, i) {
+    //let quantityInput = document.getElementsByClassName(`quantityInput`)
+    //quantityInput = quantityInput.value
+    console.log(element, i)
+}
+
+// Retrait d'une ligne produit du panier au clic sur le bouton "supprimer"
 
 let removeProductBtn = document.getElementsByClassName(`remove_product_btn`)
 const TOTALPRICE = document.getElementById(`totalPrice`)
@@ -39,19 +49,21 @@ const TOTALPRICE = document.getElementById(`totalPrice`)
 for (let i = 0; i < removeProductBtn.length; i++) {
     let button = removeProductBtn[i]
     button.addEventListener(`click`, (e) => {
-        let buttonClicked = e.target
-        buttonClicked.parentElement.parentElement.remove()
-        let product = cartContent[i]
-        //console.log(product._id)
-        function Match(id) {
-            return id._id === product._id
+        if (confirm(`Voulez-vous vraiment supprimer ce produit de votre panier ?`)) {
+            let buttonClicked = e.target
+            buttonClicked.parentElement.parentElement.remove()
+            let product = cartContent[i]
+            //console.log(product._id)
+            function Match(id) {
+                return id._id === product._id
+            }
+            let index = cartContent.findIndex(Match)
+            //console.log(index)
+            cartContent.splice(index, 1)
+            localStorage.setItem(`cart-key`, JSON.stringify(cartContent))
+
+            window.location.reload()
         }
-        let index = cartContent.findIndex(Match)
-        //console.log(index)
-        cartContent.splice(index, 1)
-        localStorage.setItem(`cart-key`, JSON.stringify(cartContent))
-        //alert(`Voulez-vous vraiment supprimer ce produit de votre panier ?`)
-        window.location.reload()
     })
 }
 
@@ -60,7 +72,7 @@ if (removeProductBtn.length === 0) {
     TOTALPRICE.innerText = `0 €`
 }
 
-// Fonction permettant de vider tout le contenu du panier eà l'écran et dans le Local Storage 
+// Fonction permettant de vider tout le contenu du panier à l'écran et dans le Local Storage 
 // lorsqu'on clique sur le bouton "Vider le panier"
 
 let clearCartBtn = document.getElementsByClassName(`empty_cart__btn`)
@@ -74,28 +86,6 @@ clearCartBtn[0].addEventListener(`click`, () => {
     document.getElementById(`emptyCartAlert`).innerHTML = emptyCartAlert
     TOTALPRICE.innerText = `0 €`
 })
-
-// changement de la quantité dans le panier
-
-/*let quantityInput = document.getElementsByClassName(`quantityInput`)
-
-for (let i = 0; i < quantityInput.length; i++) {
-    let productQuantity = quantityInput[i]
-    productQuantity.addEventListener('click', (e) => {
-        let productQuantityClicked = e.target
-        //console.log(productQuantityClicked.getAttribute(`value`))
-        let quantity = productQuantityClicked.getAttribute(`value`)
-        console.log(quantity)
-        let product = cartContent[i]
-        console.log(product._id)
-        function Match(id) {
-            return id._id === product._id
-        }
-        let index = cartContent.findIndex(Match)
-        console.log(index)
-    })
-}*/
-
 
 // Calcul et affichage du montant total du panier
 
@@ -113,8 +103,6 @@ function TotalPriceInCart() {
 }
 
 TotalPriceInCart()
-
-
 
 // Création de la requête JSON contenant l'objet contact et le tableau products 
 //lorsqu'on clique sur "Commander"
